Add ability to replace a DashCanvasView with a different view spec
