+++ desktop/cmp/dash/canvas/impl/utils.js
@@ -1,0 +1,129 @@
+/*
+ * This file belongs to Hoist, an application development toolkit
+ * developed by Extremely Heavy Industries (www.xh.io | info@xh.io)
+ *
+ * Copyright © 2022 Extremely Heavy Industries Inc.
+ */
+
+import {Icon} from '@xh/hoist/icon';
+import {wait} from '@xh/hoist/promise';
+
+/**
+ * Used to create view menu items (for adding or replacing views)
+ * @param {DashCanvasModel} dashCanvasModel
+ * @param {Object} clickPosition - {x, y}
+ * @param {string} viewIdToReplace - if replacing an existing view
+ * @returns {Object[]}
+ */
+export const createViewMenuItems = ({dashCanvasModel, clickPosition, viewIdToReplace}) => {
+    const groupedItems = {},
+        ungroupedItems = [],
+        x = clickPosition?.x ?? 0,
+        y = clickPosition?.y ?? 0,
+        addPosition = !viewIdToReplace ? calcAddPosition(x, y, dashCanvasModel) : null;
+
+    const addToGroup = (item, groupName) => {
+        const group = groupedItems[groupName];
+        if (group) {
+            group.push(item);
+        } else {
+            groupedItems[groupName] = [item];
+        }
+    };
+
+    dashCanvasModel.viewSpecs
+        .filter(viewSpec => {
+            return viewSpec.allowAdd &&
+                (
+                    !viewSpec.unique ||
+                    !(dashCanvasModel.getItemsBySpecId(viewSpec.id).length)
+                );
+        })
+        .forEach(viewSpec => {
+            const {title, icon, groupName, id} = viewSpec,
+                item = {
+                    text: title,
+                    icon: icon,
+                    actionFn: () => {
+                        if (viewIdToReplace) {
+                            dashCanvasModel.replaceView(id, viewIdToReplace);
+                        } else {
+                            const viewModel = dashCanvasModel.addView(
+                                id,
+                                {layout: addPosition}
+                            );
+                            waitFor(()=>viewModel.ensureVisible())
+                                .then(()=>viewModel.ref.current.scrollIntoView({behavior: 'smooth', block: 'center'}));
+                        }
+                    }
+                };
+
+            if (groupName) {
+                addToGroup(item, groupName);
+            } else {
+                ungroupedItems.push(item);
+            }
+        });
+
+
+    return [
+        ...Object.keys(groupedItems)
+            .map(group => {
+                // If we have any ungrouped root items (i.e. in a 'mixed mode'),
+                // insert a hidden icon to align the item text.
+                const icon = ungroupedItems.length ? Icon.angleRight({style: {visibility: 'hidden'}}) : undefined,
+                    text = group,
+                    items = groupedItems[group];
+                return {icon, text, items};
+            }),
+        ...ungroupedItems
+    ];
+};
+
+/**
+ * Used to set the {x, y} position for the next added item based on the mouse location when
+ * context menu is triggered
+ * @param {number} x - clientX position
+ * @param {number} y - clientY position
+ * @param {DashCanvasModel}
+ */
+const calcAddPosition = (x, y, dashCanvasModel) => {
+    const calcXY = (positionParams, top, left, w=0, h=0) => {
+        const calcGridColWidth = (positionParams) => {
+            const { margin, containerPadding, containerWidth, cols } = positionParams;
+            return (
+                (containerWidth - margin[0] * (cols - 1) - containerPadding[0] * 2) / cols
+            );
+        };
+
+        const clamp = (num, lowerBound, upperBound) =>
+            Math.max(Math.min(num, upperBound), lowerBound);
+
+        const {margin, cols, rowHeight, maxRows} = positionParams;
+        const colWidth = calcGridColWidth(positionParams);
+        let x = Math.round((left - margin[0]) / (colWidth + margin[0]));
+        let y = Math.round((top - margin[1]) / (rowHeight + margin[1]));
+
+        x = clamp(x, 0, cols - w);
+        y = clamp(y, 0, maxRows - h);
+        return { x, y };
+    };
+
+    const {margin, columns: cols, rowHeight, maxRows, ref, containerPadding} = dashCanvasModel,
+        containerPosition = ref.current.getBoundingClientRect(),
+        {left: containerLeft, top: containerTop, width: containerWidth} = containerPosition,
+        positionParams = {margin, cols, rowHeight, maxRows,
+            containerPadding: containerPadding ?? margin, containerWidth},
+        left = x - containerLeft,
+        top = y - containerTop;
+
+    return calcXY(positionParams, top, left);
+};
+
+const waitFor = (conditionFn, interval=1) => {
+    const check = (resolve) => {
+        if (conditionFn()) resolve();
+        else wait(interval).then(()=>check(resolve));
+    };
+    return new Promise(check);
+};
